hw_5/app_drone/requirements.txt
absl-py==0.1.10
appdirs==1.4.3
appnope==0.1.0
asn1crypto==0.23.0
autograd==1.2
backports.functools-lru-cache==1.4
bleach==1.5.0
cached-property==1.4.0
certifi==2017.11.5
cffi==1.11.2
chardet==3.0.4
conda==4.4.8
coverage==4.5.1
cpplint==1.3.0
cryptography==2.1.4
CVXcanon==0.1.1
cvxopt==1.1.9
cvxpy==0.4.11
cycler==0.10.0
Cython==0.27.3
decorator==4.2.1
dill==0.2.7.1
Django==2.0.4
ecos==2.0.5
entrypoints==0.2.3
fastcache==1.0.2
-e git+git@github.com:cathywu/learning-traffic.git@a4acec8e3d91905597ef7cf3b219af22849fbac8#egg=flow
future==0.16.0
geopy==1.12.0
googlemaps==2.5.1
gunicorn==19.7.1
gym==0.9.2
heroku==0.1.4
html5lib==0.9999999
idna==2.6
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.1
Jinja2==2.10
joblib==0.11
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
Keras==2.1.4
Lasagne==0.1
lxml==4.2.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.2
mistune==0.8.3
multiprocess==0.70.4
nbconvert==5.3.1
nbformat==4.4.0
nbopen==0.5
networkx==2.1
nose==1.3.7
nose2==0.7.4
notebook==5.4.0
numpy==1.14.0
opencv-python==3.4.0.12
ortools==6.7.4973
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
path.py==11.0.1
patsy==0.5.0
pexpect==4.4.0
pickleshare==0.7.4
Pillow==5.0.0
pipenv==11.9.0
ply==3.11
prompt-toolkit==1.0.15
protobuf==3.5.1
ptyprocess==0.5.2
pycosat==0.6.3
pycparser==2.18
pyglet==1.3.0
Pygments==2.2.0
pygrib==2.0.2
Pyomo==5.3
pyOpenSSL==17.5.0
pyparsing==2.2.0
PyPrind==2.11.2
pyproj==1.9.5.1
pypsa==0.13.0
PySocks==1.6.7
python-dateutil==1.5
pytz==2017.3
PyUtilib==5.6
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==17.0.0
quadprog==0.1.6
requests==2.18.4
-e git+https://github.com/cathywu/rllab-multiagent.git@a692db514dd8a00264e93058e2c20fbe34ce56b9#egg=rllab
ruamel-yaml==0.11.14
scikit-fmm==0.0.9
scikit-image==0.13.1
scipy==1.0.0
seaborn==0.8.1
Send2Trash==1.5.0
simplegeneric==0.8.1
simpy==3.0.10
six==1.11.0
smop==0.0.0
statsmodels==0.8.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
terminado==0.8.1
testpath==0.3.1
Theano==0.8.2
toolz==0.9.0
tornado==4.5.3
traitlets==4.3.2
urllib3==1.22
virtualenv==15.1.0
virtualenv-clone==0.3.0
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
xlrd==1.1.0
